wip: #8 Removed nextcloud user and added nginx template
An admin user is enough, user should configure its own personal users if
needed.

Also, this way we can tidy up the cli calls and let only e-mail (wich
has been renamed from username, as the "username" had to be a valid
email for let's encrypt configuration), gandi.net key and gandi.net
domain name.

Added jinja2 templating to nginx will allow us to make conditional nginx
configuration, if a user doesn't want a specific service or
service-group to be installed, we'll remove it from nginx config (as it
wont be available).

This is half-implemented, a new property has been added to main object,
containing the configured-to-be class names and a new nginx template has
been added taking it in account, still untested.