wip: Traefik integration
This removes:

- nginx
- certmanager

Also, lazylibrarian will be replaced soon by readarr (as the config file
and trackers are handled the same way on different *arrs and that
simplifies this software *a lot*). Under the same reasoning, medusa has
been replaced with sonarr.

Integrating via traefik with docker and letsencrypt made the templates
obsolte, so nothing will be template-based from now on. Just a few
static files (indexer configurations and the docker-compose itself).

As a bonus, onbi has been added (to request downloads, integrated with
radarr, lidarr and sonarr).

TODOs:

- Does not even start at this point.
- Configure onbi with sonarr, radarr, lidarr and jellyfin api keys
- I think the *arr will require the download directories to be setup
- As I'm deprecating onlyoffice, configure collabora instead.
- Web interface (altough, everything is made automatic, but with a web
  interface I could add more dns providers...)
- Document how to get the different API keys and buy domains? This is
  really the only not-so-user-friendly step